Be robust to a change in the default argument naming algorithm.
Overlay for coq/coq#12756.

This overlay preserves all the old names; the new default names
are just `t` instead of `t0`.

For some context, the old algorithm saw the unqualified name `t`
referring to (apparently) `L.V.XSet.t` and refreshed the argument
names to not conflict, which I don't believe is necessary.

The only exception to this is in `VecArith.v` on `vec_ge_mor`,
where the two sets of arguments `x y` coming from unfolding the
`Proper (... ==> ... ==> ...)` type used to be refreshed into
`x y _ x0 y0 _`, and are now by default the overlapping `x y _ x y _`.
Thus if you want to refer to the second set of arguments in an explicit
application like `vec_ge_mor (x0 := ...)`, they have to be explicitly
renamed as is done here.