Adapt to Coq/Coq#18164 (#49)
* Adapt to Coq/Coq#18164

Removal of many Coq.Arith files

* Added Util/Nat/NatCompat.v

For backwards compatibility, NatCompat.v is added. This is necessary
because Util/Nat/NatUtil.v also comes with many implicit arguments (in
particular it imports RelUtil) which are not always assumed in the rest
of the library. The lemmas from NatCompat are used with a qualified name
in order to prevent incompatibility with existing lemmas in Coq.Arith in
Coq < 8.19.

TODO: Make this library more homogeneous with respect to implicit
arguments.